Move create_client and edit_client to a new file (#148) Moved the client functions from user_and_site.py to client.py
