+++ pvsite_datamodel/write/__init__.py
@@ -2,19 +2,18 @@
 Functions for writing to the PVSite database
 """
 
+from .client import create_client, edit_client
 from .forecast import insert_forecast_values
 from .generation import insert_generation_values
 from .user_and_site import (
     add_site_to_site_group,
     change_user_site_group,
-    create_client,
     create_site,
     create_site_group,
     create_user,
     delete_site,
     delete_site_group,
     delete_user,
-    edit_client,
     edit_site,
     make_fake_site,
     update_user_site_group,
